Ref: extracting new version of depictSecurity() helper.
